LayoutTests/http/tests/websocket/tests/hybi/unpaired-surrogates-in-close-reason-expected.txt

CONSOLE MESSAGE: WebSocket close message contains invalid character(s).
Checks whether SYNTAX_ERR is thrown when attemping to close the connection with unpaired surrogates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS ws.close(1000, 'í ‡'); threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
